test(desktops): split the Desktop verbs' bridge fakes out of the suite
CI's lint on the merge with main put `DesktopCommandTests.swift`
over the §2.1 ceiling (#1346 added to the same file). The fakes
move to `DesktopBridgeFakes.swift` under a `DesktopVerb` prefix —
eight suites carry file-private twins of the same names — and
the suite aliases them file-privately, so its body, where main's
hunks sit, is untouched.

fixes #1345
